Add margin and fadeIn options to freetile.js
Add gutter option to freetile.js. Option adds a margin
to each element. Defaults to 0.

Add fadeIn with fadeInSpeed to freetile.js. Option adds
a jquery fadeIn to the tiles. Adjustable speed. Default for
fadeIn is false, default for fadeInSpeed is 100. fadeInSpeeed
is ignored if fadeIn is not set to true.

Update minified js version as well.